Refactored Pretty
